cosmetics

+++ mp-jvm/src/main/kotlin/demo/multi/helpers.kt
@@ -3,4 +3,4 @@ package demo.multi
 
 class TimeRange(val millis: Int)
 val Int.days get() = TimeRange(this * 24 * 60 * 60 * 1000)
-operator fun DateTime.plus(r: TimeRange) = DateTime(this.time.toLong() + r.millis)
+operator fun DateTime.plus(r: TimeRange) = DateTime(this.time + r.millis)
